// app/interfaces/post.ts
import { type Author } from "./author";

interface Post {
  introduction: string;
  tags: string;
  slug: string;
  title: string;
  date: string;
  coverImage: string;
  author: Author;
  excerpt: string;
  ogImage: {
    url: string;
  };
  content: string;
  preview?: boolean;
};

export default Post;